# contracts/http/belcanto-v1.openapi.yaml
openapi: 3.1.0
...
info:
  title: Belcanto API
  version: 0.4.0
  description: >-
    HTTP contract for closed B.0 onboarding, L.1 internal Lesson scheduling
    and P.1 session security (device inventory, targeted revocation, the
    account security feed and one-time password recovery).
    There is intentionally no public sign-up operation: the school creates a
    Student and the invited person only activates the prepared Account.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/activations/preview:
    post:
      summary: Preview an activation invitation
      operationId: previewActivation
      tags: [Activation]
      security: []
      description: Validates an invitation without consuming it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationPreviewRequest"
      responses:
        "200":
          description: Safe invitation preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationPreview"
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/complete:
    post:
      summary: Activate an invited account
      operationId: completeActivation
      tags: [Activation]
      security: []
      description: >-
        Consumes one invitation and lets the invited person choose the Account
        password. Success does not create a session; sign-in remains a separate
        explicit step.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteActivationRequest"
      responses:
        "204":
          description: Account activated; no session was created.
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/progress:
    post:
      summary: Read multi-step activation progress
      operationId: activationProgress
      tags: [Activation]
      security: []
      description: >-
        Server-saved activation steps for an issued invitation (AUTH-10):
        the person resumes instead of repeating confirmed steps. The token
        travels in the body so it never reaches URLs or logs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationTokenRequest"
      responses:
        "200":
          description: Current activation progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivationProgressView"
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/password:
    post:
      summary: Set the activation password (step 1)
      operationId: setActivationPassword
      tags: [Activation]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationPasswordRequest"
      responses:
        "204":
          description: Password stored; the account stays pending until finish.
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/contact:
    post:
      summary: Start contact confirmation (step 2)
      operationId: startActivationContact
      tags: [Activation]
      security: []
      description: >-
        Requires the password step first; supersedes any previous open
        confirmation and delivers a six-digit code out-of-band.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationContactRequest"
      responses:
        "204":
          description: Confirmation code issued.
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/contact/verify:
    post:
      summary: Verify the activation contact code
      operationId: verifyActivationContact
      tags: [Activation]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCodeRequest"
      responses:
        "204":
          description: Contact confirmed.
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/twofa:
    post:
      summary: Start optional 2FA enrollment (step 3)
      operationId: startActivationTwofa
      tags: [Activation]
      security: []
      description: Requires the confirmed contact step first.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationTokenRequest"
      responses:
        "200":
          description: Shared secret and otpauth provisioning URI.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwofaEnrollment"
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/twofa/confirm:
    post:
      summary: Confirm activation 2FA enrollment
      operationId: confirmActivationTwofa
      tags: [Activation]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationCodeRequest"
      responses:
        "200":
          description: One-time recovery codes, shown exactly once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodesResponse"
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/activations/finish:
    post:
      summary: Finish activation
      operationId: finishActivation
      tags: [Activation]
      security: []
      description: >-
        Requires the password and confirmed-contact steps; consumes the
        invitation, activates the account and records the verified contact.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivationFinishRequest"
      responses:
        "204":
          description: Account activated.
        "400":
          $ref: "#/components/responses/ActivationInvalid"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/sessions:
    post:
      summary: Sign in
      operationId: createSession
      tags: [Sessions]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: >-
            Either a new token pair, or a short-lived second-factor
            challenge when TOTP is enrolled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInOutcome"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/sessions/twofa:
    post:
      summary: Complete second-factor sign-in
      operationId: signInWithTwofa
      tags: [Sessions]
      security: []
      description: >-
        Exchanges an open second-factor challenge plus an authenticator or
        one-time recovery code for a session token pair. Five failed
        attempts burn the challenge.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwofaSignInRequest"
      responses:
        "200":
          description: New access and refresh token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokens"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/sessions/refresh:
    post:
      summary: Rotate a session
      operationId: refreshSession
      tags: [Sessions]
      security: []
      description: Rotates the refresh token and replaces its session record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshSessionRequest"
      responses:
        "200":
          description: Rotated access and refresh token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionTokens"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/sessions/current:
    delete:
      summary: Sign out the current session
      operationId: revokeCurrentSession
      tags: [Sessions]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Current session revoked, or already inactive.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/password-resets:
    post:
      summary: Request password recovery
      operationId: requestPasswordReset
      tags: [Security]
      security: []
      description: >-
        Starts one-time password recovery. The response never reveals
        whether the phone belongs to an account; delivery happens
        out-of-band and any previous unconsumed recovery link is
        superseded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPasswordResetRequest"
      responses:
        "204":
          description: Recovery accepted; nothing about account existence is revealed.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/password-resets/complete:
    post:
      summary: Complete password recovery
      operationId: completePasswordReset
      tags: [Security]
      security: []
      description: >-
        Consumes the one-time recovery token, rotates the credential and
        revokes every session family of the account. The token is the
        idempotency mechanism: a consumed, superseded or expired token is
        rejected as unauthenticated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompletePasswordResetRequest"
      responses:
        "204":
          description: Password rotated; all previous sessions are revoked.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/sessions:
    get:
      summary: List active sessions
      operationId: listMySessions
      tags: [Security]
      security:
        - bearerAuth: []
      description: >-
        The account's active signed-in devices, newest first, with the
        calling session marked as current.
      responses:
        "200":
          description: Active sessions of the calling account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionDevice"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/sessions/revoke-others:
    post:
      summary: Revoke all other sessions
      operationId: revokeOtherSessions
      tags: [Security]
      security:
        - bearerAuth: []
      description: >-
        Requires fresh password re-authentication and revokes every session
        family except the calling one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeSessionRequest"
      responses:
        "200":
          description: Number of revoked session families.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeOtherSessionsResult"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/sessions/{sessionId}/revoke:
    post:
      summary: Revoke one session
      operationId: revokeMySession
      tags: [Security]
      security:
        - bearerAuth: []
      description: >-
        Requires fresh password re-authentication and revokes the selected
        device's session family. The current session is refused; use
        sign-out instead.
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeSessionRequest"
      responses:
        "204":
          description: Session family revoked.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/security-events:
    get:
      summary: List account security events
      operationId: listSecurityEvents
      tags: [Security]
      security:
        - bearerAuth: []
      description: >-
        Privacy-safe projection of the account's own security history from
        the append-only audit trail, newest first, cursor-paged.
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: One page of the account's security history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityEventsPage"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/bootstrap:
    get:
      summary: Load the current account bootstrap view
      operationId: getBootstrapView
      tags: [Sessions]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Persona roles, effective permissions, and personal start data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapView"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/contacts:
    get:
      summary: List verified contacts
      operationId: listMyContacts
      tags: [Security]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Verified contacts of the calling account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerifiedContact"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/contacts/change:
    post:
      summary: Start a contact change
      operationId: startContactChange
      tags: [Security]
      security:
        - bearerAuth: []
      description: >-
        Requires fresh password re-authentication (ACC-03); supersedes any
        open confirmation and delivers a six-digit code out-of-band.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartContactChangeRequest"
      responses:
        "204":
          description: Confirmation code issued.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/contacts/confirm:
    post:
      summary: Confirm a contact change
      operationId: confirmContactChange
      tags: [Security]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmContactChangeRequest"
      responses:
        "200":
          description: The verified contact after confirmation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifiedContact"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/Internal"
  /v1/me/twofa:
    get:
      summary: Read two-factor status
      operationId: twofaStatus
      tags: [Security]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enrollment state and remaining recovery codes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwofaStatus"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
...
  /v1/me/deletion-request/cancel:
    post:
      summary: Cancel the open deletion request
      operationId: cancelDeletionRequest
      tags: [Security]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The cancelled deletion request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletionRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    InvalidInput:
      description: Request shape or value is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/InvalidInputError"
    ActivationInvalid:
      description: Invitation is invalid, expired, revoked, superseded, or consumed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ActivationInvalidError"
...
  schemas:
...
    ActivationPreviewRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 43
          maxLength: 43
          pattern: '^[A-Za-z0-9_-]{43}$'
    ActivationPreview:
      type: object
      additionalProperties: false
      required: [invitationId, kind, displayName, maskedPhone, expiresAt]
      properties:
        invitationId:
          type: string
        kind:
          type: string
          enum: [owner_bootstrap, staff_activation, student_activation]
        displayName:
          type: string
        maskedPhone:
          type: string
        expiresAt:
          type: string
          format: date-time
...